Move DefaultMetricsSourcesManager to impl submodule to break circular dependencies
