+++ ts/libs/core/src/lib/runtime/index.ts
@@ -1,2 +1,1 @@
-export * from './internal';
 export * from './public';
