--- ts/libs/core/src/lib/runtime/internal/index.ts
@@ -1,1 +1,0 @@
-export * from './metrics-source';
