--- ts/libs/core/src/lib/runtime/internal/metrics-source/index.ts
@@ -1,1 +1,0 @@
-export * from './default-metrics-sources-manager';
